% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ObservationIndicators.R
\name{add_ObservationIndicators}
\alias{add_ObservationIndicators}
\title{Add all the indicators to the lake properties table}
\usage{
add_ObservationIndicators(df_cw)
}
\arguments{
\item{df_cw}{Dataframe of the CrowdWater lake properties, prepared by
\code{\link[=prepare_data]{prepare_data()}}}
}
\value{
df_cw extended by 4 indicator columns
}
\description{
Add all the indicators to the lake properties table
}
