% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_utils.R
\name{assign_points_SC}
\alias{assign_points_SC}
\title{Assign Points of Single choice answer}
\usage{
assign_points_SC(df, property, categories, points, NA_answers = "No answer")
}
\arguments{
\item{df}{Prepared CrowdWater data frame}

\item{property}{Character Value. One of the column names of df}

\item{categories}{Character vector that lists the catetories of the property
for which points are assigned}

\item{points}{Numeric vector of the same length and order of the categories
vector with the assigned points}

\item{NA_answers}{Character vector defining one or more answers that are
treated as NA values}
}
\value{
a list of
\itemize{
\item points: Numeric vector of the category points\cr
\item valid_answer: Logical vector if all the required properties were observed\cr
\item scoring_system: The points assigned to the categories
}
}
\description{
Used within the indicator functions
}
