% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_utils.R
\name{base_map}
\alias{base_map}
\title{the Basemap used for all leaflet maps}
\usage{
base_map(
  longitude = 13.3987,
  latitude = 52.5048,
  zoom = 11,
  berlin_districts = TRUE
)
}
\arguments{
\item{longitude}{Numeric value defining the longitude if WGS84}

\item{latitude}{Numeric value defining the latitude if WGS84}

\item{zoom}{The initial and minimum zoom level}

\item{berlin_districts}{Logical value if berlin districts should be added}
}
\value{
A leaflet map
}
\description{
the Basemap used for all leaflet maps
}
