% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_properties_names.R
\docType{data}
\name{bboxes}
\alias{bboxes}
\title{Regions Bounding Boxes}
\format{
A named list of regions defined by bounding boxes. The bounding boxes
are  numeric vectors of coordinates which define the top, bottom, left, and
right coordinates of the box.(in WGS84)
}
\usage{
bboxes
}
\description{
Regions Bounding Boxes
}
\keyword{datasets}
