% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_properties_names.R
\docType{data}
\name{berlin_district_shp}
\alias{berlin_district_shp}
\title{Simple feature of the Berlin district boundaries}
\format{
Besides the geometry, the Name of the district and its ID is given
}
\usage{
berlin_district_shp
}
\description{
Simple feature of the Berlin district boundaries
}
\keyword{datasets}
