% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_sum.R
\name{categories_sum}
\alias{categories_sum}
\title{This function returns the sum of categories per property of the given
dataframe}
\usage{
categories_sum(df, property, language = "english", includeNA = FALSE)
}
\arguments{
\item{df}{Dataframe of the crowdwater lake properties prepared by
\code{\link[=prepare_data]{prepare_data()}}, optionally filtered for one ROOT_ID}

\item{property}{The property}

\item{language}{A character string, defining the language of df. If unknown
the function \code{\link[=identify_language]{identify_language()}} can be used to find out.}

\item{includeNA}{Logical if NA (No answer) data should be included}
}
\value{
A dataframe with the columns ROOT_ID, LATITUED, LONGITUDE and all
the available categories per defined property. Besides the percentage of
categories per site, the number of observations is given.
}
\description{
This function returns the sum of categories per property of the given
dataframe
}
