% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_utils.R
\name{create_legend_labels}
\alias{create_legend_labels}
\title{Adds additional information to legend values}
\usage{
create_legend_labels(
  lv,
  low_label = NULL,
  high_label = NULL,
  increase_bounds = TRUE
)
}
\arguments{
\item{lv}{legend values as returned by \code{\link[=indicators_to_colors]{indicators_to_colors()}} in the
legend_table}

\item{low_label}{Additional lable for the lowest value}

\item{high_label}{Additional lable for the highest value}

\item{increase_bounds}{Add greater or equal / less or equal than sign to
the highest /lowest legend value}
}
\value{
extended legend labels
}
\description{
Adds additional information to legend values
}
