% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LakeIndicators.R
\name{get_LakeIndicators}
\alias{get_LakeIndicators}
\title{Calculate all indicators for all recorded lakes}
\usage{
get_LakeIndicators(df_cw)
}
\arguments{
\item{df_cw}{Dataframe of the crowdwater lake properties prepared by
\code{\link[=prepare_data]{prepare_data()}}}
}
\value{
A dataframe that contains columns of the root id, coordinates,
the indicator values and number of observations as well as the date of the
most recent observation
}
\description{
Calculate all indicators for all recorded lakes
}
