% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identify_language}
\alias{identify_language}
\title{Function identifies the language based on category names}
\usage{
identify_language(df_cw)
}
\arguments{
\item{df_cw}{Dataframe of the crowdwater lake properties prepared by
\code{\link[=prepare_data]{prepare_data()}}}
}
\value{
A character string of the language
}
\description{
Function identifies the language based on category names
}
