% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_nutrients.R
\name{indicator_nutrients}
\alias{indicator_nutrients}
\title{Calculation of the CrowdWater Nutrient Indicator}
\usage{
indicator_nutrients(df_cw, return_scoring_system = FALSE)
}
\arguments{
\item{df_cw}{Dataframe of the prepared crowdwater data}

\item{return_scoring_system}{If TRUE the scoring system is returned instead
of the indicator vector.}
}
\value{
Either a numeric vector of the indicator points, or a description of the
scoring system. This is a list of properties and categories which contribute
to the indicator.
}
\description{
The indicator is calculated per observation
}
