% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_water-stress.R
\name{indicator_waterStress}
\alias{indicator_waterStress}
\title{Calculation of the CrowdWater Biotope Indicator}
\usage{
indicator_waterStress(df_cw, return_scoring_system = FALSE)
}
\arguments{
\item{df_cw}{Dataframe of the crowdwater lake properties prepared by
\code{\link[=prepare_data]{prepare_data()}}}

\item{return_scoring_system}{If TRUE the scoring system is returned instead
of the indicator vector.}
}
\value{
Either a numeric vector of the indicator points, or a description of the
scoring system. This is a list of properties and categories which contribute
to the indicator.
}
\description{
The indicator is calculated per observation
}
\details{
While most citizens are not able to tell if the lake experiences water stress
by the observation at one point of time, there might be citizens who know the
water body for a long time. The more observations include this information
the more certain it is. Some citizen may only know about water level changes
but cannot answer if the lake dries out. This is why in contrast to the other
indicators, the points are counted even if one question remains unknown or
unanswered.
}
