% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_utils.R
\name{indicators_to_colors}
\alias{indicators_to_colors}
\title{Transformation of numeric values into colors}
\usage{
indicators_to_colors(
  v_indicator,
  v_worst,
  v_neutral,
  v_best,
  col_worst = "#E6550D",
  col_neutral = "#EFF3FB",
  col_best = "#08306B",
  col_NA = "#00000000",
  resolution_bad = 0.1,
  resolution_good = 0.1,
  resolution_legend = 1
)
}
\arguments{
\item{v_indicator}{Numeric Vector of indicator data}

\item{v_worst}{Worst possible value}

\item{v_neutral}{Neutral value}

\item{v_best}{Best possible value}

\item{col_worst}{Color for worst value as hexadecimal string of the
form "#rrggbb"}

\item{col_neutral}{Color for worst value as hexadecimal string of the
form "#rrggbb"}

\item{col_best}{Color for worst value as hexadecimal string of the
form "#rrggbb"}

\item{col_NA}{Color for NA values (Transparent by default)}

\item{resolution_bad}{the resolution between neutral and worst value}

\item{resolution_good}{The resolution between netral and best value}

\item{resolution_legend}{The resolution of the legend values}
}
\value{
Vector of colors corresponding to the indicator values
}
\description{
Transformation of numeric values into colors
}
