% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_indicator.R
\name{lake_indicator}
\alias{lake_indicator}
\title{From single observation to average site indicators}
\usage{
lake_indicator(df_cw, indicator_type)
}
\arguments{
\item{df_cw}{Dataframe of the crowdwater lake properties prepared by
\code{\link[=prepare_data]{prepare_data()}}}

\item{indicator_type}{A character defining the indicator. Either "nutrients"
"biotope", "usage" or "waterstress"}
}
\value{
A data frame of 5 columns (ROOT_ID, LATITUDE, LONGITUDE and the indicator
value and the number of observations)
}
\description{
From single observation to average site indicators
}
