% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_observations.R
\name{latest_observations}
\alias{latest_observations}
\title{Returnes only the latest observations per lake}
\usage{
latest_observations(df_cw)
}
\arguments{
\item{df_cw}{Dataframe of the crowdwater lake properties prepared by
\code{\link[=prepare_data]{prepare_data()}}}
}
\value{
A dataframe in the same shape as the lake properties dataframe with
all historical observations removed
}
\description{
Returnes only the latest observations per lake
}
