% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_indicators.R
\name{leaflet_indicator}
\alias{leaflet_indicator}
\title{Create an interactive leaflet map which contains all the information of the
last observation}
\usage{
leaflet_indicator(cw_indicators, language = "german")
}
\arguments{
\item{cw_indicators}{Dataframe of the crowdwater indicators per lake created
by \code{\link[=get_LakeIndicators]{get_LakeIndicators()}}}

\item{language}{A character string, defining the language if the map}
}
\value{
A leflet map
}
\description{
The leaflet map has two layers. 1) all the observered properties
2) the corresponding image of that observation
}
