% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_last_observation.R
\name{leaflet_last_observation_days}
\alias{leaflet_last_observation_days}
\title{Create an interactive leaflet map which contains the days passed since the
last observations}
\usage{
leaflet_last_observation_days(df_cw, language = "german")
}
\arguments{
\item{df_cw}{Dataframe of the crowdwater lake properties prepared by
\code{\link[=prepare_data]{prepare_data()}}}

\item{language}{A character string, defining the language if the map}
}
\value{
A leaflet map
}
\description{
Create an interactive leaflet map which contains the days passed since the
last observations
}
