% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_last_observation.R
\name{leaflet_last_observation_info}
\alias{leaflet_last_observation_info}
\title{Create an interactive leaflet map which contains all the information of the
last observation}
\usage{
leaflet_last_observation_info(df_cw, language = "german")
}
\arguments{
\item{df_cw}{Dataframe of the crowdwater lake properties prepared by
\code{\link[=prepare_data]{prepare_data()}}}

\item{language}{A character string, defining the language if the map}
}
\value{
A leflet map
}
\description{
The leaflet map has two layers. 1) all the observered properties
2) the corresponding image of that observation
}
