% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LakeIndicators.R
\name{merge_LakeIndicators}
\alias{merge_LakeIndicators}
\title{Merge two lake indicator tables}
\usage{
merge_LakeIndicators(x, y)
}
\arguments{
\item{x}{indicator table created by \code{\link[=lake_indicator]{lake_indicator()}}}

\item{y}{indicator table created by \code{\link[=lake_indicator]{lake_indicator()}}}
}
\value{
merged table
}
\description{
Merge two lake indicator tables
}
