% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{CrowdWater lake data preparation}
\usage{
prepare_data(path, file, region = NULL, validated_only = TRUE)
}
\arguments{
\item{path}{The filepath}

\item{file}{The filename including the ".csv" ending}

\item{region}{Either a character defining a region which is available in
or a numeric vector of coordinates in WGS84 shaping a rectangle by "top",
"bottom", "left" and "right" coordinate.}

\item{validated_only}{Logical if only checked observations are selected}
}
\value{
A list of two data frames. The first one contains data on the observered
proberties, the second data frame contains data of the phyiscale category.
}
\description{
The downloaded csv file is filtered for lake specific columns, for regions
and for validated observations. The observed properties are named correctly
in English.
}
