% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_properties_names.R
\docType{data}
\name{props_and_cats}
\alias{props_and_cats}
\title{List of CrowdWater Properties and Categories}
\format{
A named list where list names correspond to the crowdwater properties
(column names of the table). For each property a data frame of categories is
given, which includes the translation into German or English, the order of
categories if applicable (if not -> NA) and the type of answers
}
\usage{
props_and_cats
}
\description{
List of CrowdWater Properties and Categories
}
\keyword{datasets}
