% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{renameMultipleChoice}
\alias{renameMultipleChoice}
\title{Replaces multiple separated category names by new names}
\usage{
renameMultipleChoice(v, oldNames, newNames, sep = ", ")
}
\arguments{
\item{v}{Character vector of categories}

\item{oldNames}{Character vector listing the old names of the categories}

\item{newNames}{Characteri vector listing the new names of the categories in
the same order as the oldNames vector}

\item{sep}{The seperator that is used in a cell between categories of the same
observation}
}
\value{
Character vector of renamed categories
}
\description{
Multiple choice answers per observation are combined a one cell of the table
and seperated by a comma.
}
