% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{renameSingleChoice}
\alias{renameSingleChoice}
\title{Replaces multiple category names by new names}
\usage{
renameSingleChoice(v, oldNames, newNames)
}
\arguments{
\item{v}{Character vector of categories}

\item{oldNames}{Character vector listing the old names of the categories}

\item{newNames}{Characteri vector listing the new names of the categories in
the same order as the oldNames vector}
}
\value{
Character vector of renamed categories
}
\description{
Replaces multiple category names by new names
}
