% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_category_percentage.R
\name{site_category_percentage}
\alias{site_category_percentage}
\title{Returnes only the latest observations per lake}
\usage{
site_category_percentage(df_cw, includeNA = FALSE)
}
\arguments{
\item{df_cw}{Dataframe of the crowdwater lake properties prepared by
\code{\link[=prepare_data]{prepare_data()}}}

\item{includeNA}{Logical if NA (No answer) data should be included}
}
\value{
A dataframe in the same shape as the lake properties dataframe with
all historical observations removed
}
\description{
Returnes only the latest observations per lake
}
