% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_categories.R
\name{tranlsate_categories}
\alias{tranlsate_categories}
\title{From single observation to average site indicators}
\usage{
tranlsate_categories(df_cw, from, to)
}
\arguments{
\item{df_cw}{Dataframe of the crowdwater lake properties prepared by
\code{\link[=prepare_data]{prepare_data()}}}

\item{from}{Character string, defining the language of df_cw}

\item{to}{Character string, defining the language to translate in}
}
\value{
df_cw translated
}
\description{
From single observation to average site indicators
}
